Reduce diffs and unify MSBuild forwarding
